renamed variables and moved class
