+++ reactions/gremlin/gremlin-reaction/gremlin-reaction.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>gremlin_reaction</RootNamespace>
+    <RootNamespace>Drasi.Reactions.Gremlin</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
